requirements.txt
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.2.0
ccxt>=2.0.0  # For cryptocurrency trading
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
ta>=0.10.0  # Technical analysis library
backtrader>=1.9.76.123  # For backtesting
dash>=2.0.0  # For interactive dashboards
plotly>=5.3.0  # For interactive visualizations
pytest>=6.2.5  # For testing 